requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-docx==1.1.0
reportlab==4.0.7
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==3.0.1
PyMuPDF==1.23.8
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99
